PageSchema.methods.revealSecret - Make recursive
Takes a page. If that page belongs to a character, reveals the secret to that character. If it's a category, runs itself on each of the category's members. This way, you can reveal a secret to all the characters in a category. This will also dig down through subcategories, and reveal the secret to all characters within that tree, however deep they might be.