Test utility method: createTestDocs - Create page last
We were creating the page and then creating the characters. This meant that if the test page included as secret, that secret wouldn't be revealed to the test user's character, because the test user didn't have a character yet when hen created the page. By making the page last, we make sure that the character knows about any potential secret, which avoids some unexpected test results.