Page model - Add checked to secrets
Some secrets have game-related checks inside them. The first time a character requests a page with such a secret, the game module can check to see if it's something the character knows. This keeps track of which characters have checked, so that they don't get to run the check over and over again.