Page model: Categories: Separate types for cleaner schema
We were storing a "secret" that we _said_ was a String, but in actual fact, it was sometimes a string and sometimes a boolean. We should break these into two fields: a boolean that tells us *if* the categorization is a secret, and a string which tells us the codename of the secret if it is.